refactor: Use TypeScript
