+++ eslint.config.js
@@ -1,6 +1,7 @@
 import js from '@eslint/js'
 import vue from 'eslint-plugin-vue'
 import tsParser from '@typescript-eslint/parser'
+import tsPlugin from '@typescript-eslint/eslint-plugin'
 
 export default [
     {
@@ -9,18 +10,24 @@ export default [
     js.configs.recommended,
     ...vue.configs['flat/essential'],
     {
-        files: ['**/*.{js,vue}'],
+        files: ['**/*.{js,vue,ts}'],
         languageOptions: {
             ecmaVersion: 'latest',
             sourceType: 'module',
-        },
+        }
     },
     {
-        files: ['**/*.ts'],
+        files: ['**/*.{ts,vue}'],
         languageOptions: {
             parser: tsParser,
             ecmaVersion: 'latest',
             sourceType: 'module',
+        },
+        plugins: {
+            '@typescript-eslint': tsPlugin,
+        },
+        rules: {
+            ...tsPlugin.configs.recommended.rules,
         },
     },
 ]
