+++ src/App.vue
@@ -1,4 +1,4 @@
-<script setup>
+<script setup lang="ts">
 import './styles/reset.less'
 import Wrapper from './components/Wrapper.vue'
 import Toaster from './components/widgets/Toaster.vue'
