+++ src/components/Footer.vue
@@ -1,4 +1,4 @@
-<script setup>
+<script setup lang="ts">
 defineOptions({ name: 'MainFooter' })
 </script>
