+++ src/components/widgets/Button.vue
@@ -1,4 +1,4 @@
-<script setup>
+<script setup lang="ts">
 defineOptions({ name: 'WidgetButton' })
 </script>
