--- src/env.d.ts
@@ -1,12 +1,0 @@
-/// <reference types="vite/client" />
-
-declare module '*.vue' {
-    import type { DefineComponent } from 'vue'
-    const component: DefineComponent<{}, {}, any>
-    export default component
-}
-
-declare module '*.svg' {
-    const src: string
-    export default src
-}
