+++ src/types/about.ts
@@ -1,0 +1,6 @@
+export interface AboutItem {
+    label: string
+    icon: string
+    value: string
+    link?: string
+}
