+++ src/types/env.d.ts
@@ -1,0 +1,15 @@
+/// <reference types="vite/client" />
+
+declare module '*.svg' {
+    const src: string
+    export default src
+}
+
+interface ImportMetaEnv {
+    readonly VITE_API_URL: string
+}
+
+interface ImportMeta {
+    readonly env: ImportMetaEnv
+}
+
