+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './data'
+export * from './store'
+export * from './about'
