+++ src/utils/initDetailData.ts
@@ -1,91 +1,9 @@
+/**
+ * 此 util 用于初始化详情页面数据
+ */
+
 import { setTitle } from './title'
-
-// 接口数据结构 BEGIN
-// Detail
-export interface DetailBuild {
-    number: string
-    branch: string
-    compileTime: string
-    arch: string[]
-    counterpart: string
-}
-export interface DetailRelease {
-    channel: string
-    time: string
-    url: string
-    announcePlace: string
-}
-export interface DetailFeatureIds {
-    url: string
-    fileName: string
-}
-export interface DetailNavRef {
-    category: string
-    build: string
-}
-export interface DetailBelongsTo {
-    path: string
-    name: string
-}
-export interface DetailUpdateId {
-    arch: string
-    id: string
-}
-export interface DetailDownload {
-    name: string
-    link: { url: string; source: string }[]
-    arch: string
-    md5?: string
-    sha256?: string
-    size?: string
-}
-export interface DetailContent {
-    build: DetailBuild
-    release?: DetailRelease
-    featureIds?: DetailFeatureIds
-    nav: { previous?: DetailNavRef; next?: DetailNavRef }
-    belongsTo: DetailBelongsTo
-    updateId?: DetailUpdateId[]
-    download?: DetailDownload
-}
-
-// CategoryList
-export interface CategoryNavRef {
-    path: string
-    name: string
-}
-export interface CategoryContent {
-    name: string
-    codename: string
-    belonging: string
-    semester: string
-    range: [string, string]
-    nav: { previous?: CategoryNavRef; next?: CategoryNavRef }
-    list: [string, string][]
-}
-
-// 其它通用结构
-export interface NavItem {
-    platform: string
-    route: string
-    build?: string
-    path?: string
-}
-export interface NavData {
-    type: 'detail' | 'categoryList'
-    prev?: NavItem
-    next?: NavItem
-}
-export interface PageData {
-    data: DetailContent | CategoryContent
-    isLoading: boolean
-    nav?: NavData | null
-}
-export interface RespOrigin {
-    dataType: 'detail' | 'categoryList'
-    content: DetailContent | CategoryContent
-}
-// 结构 END
+import type { DetailContent, CategoryContent, NavData, PageData, RespOrigin } from '@/types'
 
 const buildDetailNav = (resp: DetailContent): NavData => {
     const { previous, next } = resp.nav
