+++ src/utils/parseRepoInfo.ts
@@ -1,10 +1,10 @@
-'use strict'
 /**
- * 此脚本用于在处理本项目相关信息
+ * 此 util 用于处理本项目相关信息
  */
 
 import packageInfo from '../../package.json'
 import buildInfo from '../../scripts/buildInfo.json'
+import type { AboutItem } from '@/types'
 
 // 项目信息相关
 const pkgVersion = 'v' + packageInfo.version + ' (build ' + buildInfo.build + ')' // 项目版本号
@@ -13,7 +13,7 @@ const pkgRepoName = pkgRepo.split('https://github.com/')[1] // 项目名
 
 // 获取 build tag
 const buildTag = (() => {
-    // 处理版本号和分支 
+    // 处理版本号和分支
     const buildNum = buildInfo.build
     const buildBranch = buildInfo.branch
     // 处理构建时间
@@ -30,27 +30,19 @@ const buildTag = (() => {
 })()
 
 const styleVerName = `font-size: 14px;
-                    color: #000; 
-                    background-color: #8ad5b3; 
+                    color: #000;
+                    background-color: #8ad5b3;
                     padding: 5px 9px;
                     border-radius: 3px 0 0 3px;
                     margin: 10px 0; `
-const styleVerValue = `font-size: 14px; 
-                     color: #fff; 
-                     background-color: #607d8b; 
+const styleVerValue = `font-size: 14px;
+                     color: #fff;
+                     background-color: #607d8b;
                      padding: 5px 9px;
                      border-radius: 0 3px 3px 0;
                      margin: 6px 0; `
 console.log('%c%s%c%s', styleVerName, 'wutd', styleVerValue, 'v' + packageInfo.version)
 console.log(`Build Tag: ${ buildTag } (g${ buildInfo.hash })`)
-
-// 导出数据
-export interface AboutItem {
-    label: string
-    icon: string
-    value: string
-    link?: string
-}
 
 const aboutInfo: AboutItem[] = [
     { label: '站点版本', icon: 'search', value: pkgVersion },
