+++ src/views/MainHome.vue
@@ -1,4 +1,4 @@
-<script setup>
+<script setup lang="ts">
 import { storeToRefs } from 'pinia'
 
 import icons from '@/assets/icons'
