+++ src/views/NotFound.vue
@@ -1,4 +1,4 @@
-<script setup>
+<script setup lang="ts">
 import { useRouter } from 'vue-router'
 const router = useRouter()
